Extract query functions to query module
