+++ src/lib.rs
@@ -1,5 +1,8 @@
 extern crate chrono;
 extern crate serde_derive;
+#[macro_use]
+extern crate log;
 
 pub mod hydra;
 pub mod pretty;
+pub mod query;
